// api/src/modules/logs/enums/logsEvents.ts
enum LogsEvents {
    TRANSACCION_CREADA_CORRECTAMENTE = "Transacción Creada correctamente.",
    TRANSACCION_PAGO_CREADA_CORRECTAMENTE = "Transacción de Pago Creada correctamente.",
    TRANSACCION_PAGO_ACTUALIZADA_CORRECTAMENTE = "Transacción de Pago actualizada correctamente.",
    TRANSACCION_PAGO_FINALIZADA_SALESFORCE = "Transacción de Pago Finalizada en Salesforce.",
    TRANSACCION_INFORMADA = "Transacción Informada.",
    TRANSACCION_INICIALIZADA = "Transacción Inicializada.",
    ERROR_EN_LA_RESPUESTA_DEL_SERVICIO = "Error en la respuesta del servicio.",
    LLAMADO_DE_BANCO_ACEPTADO = "Llamado de Banco autorizado recibido correctamente.",
    TRANSACCION_PENDIENTE_DE_CONFIRMACION = "Transacción pendiente de confirmación.",
    TRANSACCION_CONFIRMADA = "Transacción confirmada.",
    CREACION_DE_CLIENTE_EN_SALESFORCE = "Creación de Cliente en Salesforce.",
    CLIENTE_CREADO_EN_SALESFORCE = "Cliente creado correctamente en Salesforce.",
    LOG_INPUT = "Log Input",
  }
  export default LogsEvents;