// api/src/modules/transaccion/enums/transaccion.estados.enum.ts
export enum estadosTransaccion {
    INICIALIZADA = "INICIALIZADA",
    CREADA= "CREADA",
    CONFIRMADA_PAYU = "CONFIRMADA_PAYU",
    CONFIRMACION_PENDIENTE = "CONFIRMACION_PENDIENTE",
    CONFIRMADA_SF = "CONFIRMADA_SF",
    FINALIZADA_SF_OK = "FINALIZADA_SF_OK",
    FINALIZADA_SF_ERROR = "FINALIZADA_SF_ERROR",
    ERROR = "ERROR",
    ERROR_SF = "ERROR_SF",
}